usering normal function instead of endpoint
